% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_api.R
\name{cppMVNmixPMLE}
\alias{cppMVNmixPMLE}
\title{Low-Level C++ EM Engine for PMLE}
\description{
Interface to the underlying C++ implementation of the
penalized EM iterations. This function is exported for backward compatibility.
}
