% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnMix-package.R
\docType{package}
\name{mvnMix-package}
\alias{mvnMix}
\alias{mvnMix-package}
\title{mvnMix: EM Test for Multivariate Normal Mixtures}
\description{
Estimation and testing for finite mixtures of multivariate normal distributions. Implements the penalized maximum likelihood estimator (PMLE) and the modified EM test of Kasahara and Shimotsu for testing the number of components.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kshimotsu/mvnMix}
  \item Report bugs at \url{https://github.com/kshimotsu/mvnMix/issues}
}

}
\author{
\strong{Maintainer}: Hiroyuki Kasahara \email{hkasahar@mail.ubc.ca}

Authors:
\itemize{
  \item Katsumi Shimotsu
}

Other contributors:
\itemize{
  \item Seojeong Ahn [contributor]
}

}
\keyword{internal}
