% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_tests.R
\name{mvnmixCritBoot}
\alias{mvnmixCritBoot}
\title{mvnmixCritBoot}
\usage{
mvnmixCritBoot(
  y,
  an = 1,
  parlist,
  values = NULL,
  ninits = 10,
  nbtsp = 199,
  parallel = 0.75,
  cl = NULL,
  LRT.penalized = FALSE
)
}
\arguments{
\item{y}{n by d matrix of data}

\item{an}{Penalty-weight parameter used in PMLE updates.}

\item{parlist}{Parameter list with \code{alpha}, \code{mu}, and \code{sigma}.}

\item{values}{Optional vector of observed statistics for p-value calculation.}

\item{ninits}{The number of initial candidates to be generated.}

\item{nbtsp}{The number of bootstrap observations.}

\item{parallel}{Determines what percentage of available cores are used.}

\item{cl}{Cluster used for parallelization (optional).}

\item{LRT.penalized}{Whether penalized log-likelihood is used in bootstrap
test-statistic updates.}
}
\value{
A list with elements \code{crit} and \code{pvals}.
}
\description{
Computes bootstrap critical values for the modified EM test.
}
\details{
Compute Bootstrap Critical Values for MEM Test
}
