% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_tests.R
\name{mvnmixMEMtest}
\alias{mvnmixMEMtest}
\title{mvnmixMEMtest}
\usage{
mvnmixMEMtest(
  y,
  m = 2,
  an = 1,
  tauset = c(0.1, 0.3, 0.5),
  ninits = 10,
  crit.method = c("asy", "boot", "none"),
  nbtsp = 199,
  cl = NULL,
  parallel = 0.75,
  LRT.penalized = FALSE,
  nrep = 10000,
  ninits.crit = 25
)
}
\arguments{
\item{y}{n by d matrix of data}

\item{m}{The number of components in the mixture defined by a null hypothesis, m_0}

\item{an}{a term used for penalty function}

\item{tauset}{A set of initial tau value candidates}

\item{ninits}{The number of randomly drawn initial values.}

\item{crit.method}{Method used to compute critical values: one of
\code{"none"}, \code{"asy"}, and \code{"boot"}.}

\item{nbtsp}{The number of bootstrap observations.}

\item{cl}{Cluster used for parallelization.}

\item{parallel}{Share of available cores used for bootstrap simulation.}

\item{LRT.penalized}{Determines whether penalized likelihood is used in the
alternative model objective.}

\item{nrep}{Number of simulation draws used in asymptotic critical values.}

\item{ninits.crit}{Number of random starts used in cone-projection steps.}
}
\value{
A list containing \code{emstat}, \code{pvals}, \code{crit}, fitted null-model
outputs, and metadata.
}
\description{
Perform Modified EM Test for Mixture Order
}
\examples{
data(faithful)
y <- as.matrix(faithful[, c("eruptions", "waiting")])
mvnmixMEMtest(y = y, m = 1, crit.method = "none", ninits = 3)
}
