% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_funcs.R
\name{mvnmixPMLE}
\alias{mvnmixPMLE}
\title{mvnmixPMLE}
\usage{
mvnmixPMLE(
  y,
  m = 2,
  ninits = 100,
  epsilon = 1e-08,
  maxit = 2000,
  epsilon.short = 0.01,
  maxit.short = 500,
  binit = NULL
)
}
\arguments{
\item{y}{n by d matrix of data}

\item{m}{The number of components in the mixture}

\item{ninits}{The number of randomly drawn initial values.}

\item{epsilon}{The convergence criterion. Convergence is declared when the penalized log-likelihood increases by less than \code{epsilon}.}

\item{maxit}{The maximum number of iterations.}

\item{epsilon.short}{The convergence criterion in short EM. Convergence is declared when the penalized log-likelihood increases by less than \code{epsilon.short}.}

\item{maxit.short}{The maximum number of iterations in short EM.}

\item{binit}{The initial value of parameter vector that is included as a candidate parameter vector}
}
\value{
A list containing estimated \code{coefficients}, \code{parlist}, \code{loglik},
\code{penloglik}, \code{aic}, \code{bic}, \code{postprobs}, \code{call}, and \code{m}.
}
\description{
Estimates parameters of a finite mixture of multivariate normals
by penalized maximum log-likelhood functions.
}
\details{
Estimate PMLE for Multivariate Normal Mixtures
}
\note{
\code{mvnmixPMLE} maximizes the penalized log-likelihood function
using the EM algorithm with combining short and long runs of EM steps as in Biernacki et al. (2003).
\code{mvnmixPMLE} first runs the EM algorithm from \code{ninits}\eqn{* 4m(1 + p)} initial values
with the convertence criterion \code{epsilon.short} and \code{maxit.short}.
Then, \code{mvnmixPMLE} uses \code{ninits} best initial values to run the EM algorithm
with the convertence criterion \code{epsilon} and \code{maxit}.
}
\references{
Alexandrovich, G. (2014)
A Note on the Article `Inference for Multivariate Normal Mixtures' by J. Chen and X. Tan
\emph{Journal of Multivariate Analysis}, \bold{129}, 245--248.

Biernacki, C., Celeux, G. and Govaert, G. (2003)
Choosing Starting Values for the EM Algorithm for Getting the
Highest Likelihood in Multivariate Gaussian Mixture Models,
\emph{Computational Statistics and Data Analysis}, \bold{41}, 561--575.

Boldea, O. and Magnus, J. R. (2009)
Maximum Likelihood Estimation of the Multivariate Normal Mixture Model,
\emph{Journal of the American Statistical Association},
\bold{104}, 1539--1549.

Chen, J. and Tan, X. (2009)
Inference for Multivariate Normal Mixtures,
\emph{Journal of Multivariate Analysis}, \bold{100}, 1367--1383.

McLachlan, G. J. and Peel, D. (2000) \emph{Finite Mixture Models}, John Wiley and Sons, Inc.
}
