% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_funcs.R
\name{mvnmixPMLEinit}
\alias{mvnmixPMLEinit}
\title{mvnmixPMLEinit}
\usage{
mvnmixPMLEinit(y, ninits = 1, m = 2)
}
\arguments{
\item{y}{n by d matrix of data}

\item{ninits}{number of initial values to be generated}

\item{m}{The number of components in the mixture}
}
\value{
A list with elements \code{alpha}, \code{mu}, and \code{sigma}.
}
\description{
Generate initial values used by PMLE of multivariate normal
mixtures.
}
\details{
Generate Initial Values for PMLE
}
