% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_methods.R
\name{rmvnmix}
\alias{rmvnmix}
\title{rmvnmix}
\usage{
rmvnmix(n, alpha, mu, sigma)
}
\arguments{
\item{n}{Number of observations.}

\item{alpha}{Length-\eqn{m} vector of component probabilities.}

\item{mu}{\eqn{d \times m} matrix whose \eqn{j}-th column is component mean.}

\item{sigma}{\eqn{d \times (d m)} block matrix of component covariance matrices.}
}
\value{
An \eqn{n \times d} matrix of random draws.
}
\description{
Draw random observations from a finite mixture of multivariate
normal components.
}
\details{
Generate Random Draws from a Multivariate Normal Mixture
}
