% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_crit.R
\name{mvnmixCrit}
\alias{mvnmixCrit}
\title{mvnmixCrit}
\usage{
mvnmixCrit(y, parlist, values = NULL, nrep = 10000, ninits.crit = 25)
}
\arguments{
\item{y}{n by d matrix of data}

\item{parlist}{The parameter estimates as a list containing alpha, mu, and sigma
in the form of (alpha = (alpha_1,...,alpha_m), mu = (mu_1',...,mu_m'),
sigma = (vech(sigma_1)',...,vech(sigma_m)')}

\item{values}{Vector of test statistic values at which p-values are computed}

\item{nrep}{Number of simulation replications (default 10000)}

\item{ninits.crit}{Number of random starts for cone projection optimization (default 25)}
}
\value{
A list with:
\item{crit}{Critical values at 10\%, 5\%, 1\% levels}
\item{pvals}{P-values corresponding to \code{values}}
}
\description{
Asymptotic critical values for multivariate normal mixture MEM test
}
\details{
Computes asymptotic critical values and p-values for the modified EM test
of Kasahara and Shimotsu for multivariate normal mixtures.
}
