% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_funcs.R
\name{mvnmixMaxPhi}
\alias{mvnmixMaxPhi}
\title{mvnmixMaxPhi}
\usage{
mvnmixMaxPhi(y, parlist, an, tauset = c(0.1, 0.3, 0.5), ninits = 10,
  epsilon.short = 0.01, epsilon = 1e-08, maxit.short = 500,
  maxit = 2000, verb = FALSE, parallel = 0.75, cl = NULL)
}
\arguments{
\item{y}{n by d matrix of data}

\item{parlist}{The parameter estimates as a list containing alpha, mu, and sigma
in the form of (alpha = (alpha_1,...,alpha_m), mu = (mu_1',...,mu_m'),
sigma = (vech(sigma_1)',...,vech(sigma_m)')}

\item{an}{a term used for penalty function}

\item{tauset}{A set of initial tau value candidates}

\item{ninits}{The number of randomly drawn initial values.}

\item{epsilon.short}{The convergence criterion in short EM. Convergence is declared when the penalized log-likelihood increases by less than \code{epsilon.short}.}

\item{epsilon}{The convergence criterion. Convergence is declared when the penalized log-likelihood increases by less than \code{epsilon}.}

\item{maxit.short}{The maximum number of iterations in short EM.}

\item{maxit}{The maximum number of iterations.}

\item{verb}{Determines whether to print a message if an error occurs.}

\item{parallel}{Determines what percentage of available cores are used, represented by a double in [0,1]. 0.75 is default.}

\item{cl}{Cluster used for parallelization; if it is \code{NULL}, the system will automatically
create a new one for computation accordingly.}
}
\value{
A list with items:
\item{loglik}{Log-likelihood resulting from MEM algorithm at k=1,2,3.}
\item{penloglik}{Penalized log-likelihood resulting from MEM algorithm at k=1,2,3.}
}
\description{
Compute ordinary & penalized log-likelihood ratio resulting from
MEM algorithm at k=1,2,3.
}
