% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_funcs.R
\name{mvnmixPMLEinit}
\alias{mvnmixPMLEinit}
\title{mvnmixPMLEinit}
\usage{
mvnmixPMLEinit(y, ninits = 1, m = 2)
}
\arguments{
\item{y}{n by d matrix of data}

\item{ninits}{number of initial values to be generated}

\item{m}{The number of components in the mixture}
}
\value{
A list with the following items:
\item{alpha}{m by ninits matrix for alpha}
\item{mu}{d*m by ninits matrix for mu, each column is (mu_1',...,mu_m')'}
\item{sigma}{d(d+1)/2*m by ninits matrix for sigma, each column is (vech(sigma_1)',...,vech(sigma_m)')'}
}
\description{
Generate initial values used by the PMLE of multivariate normal mixture
}
