% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_methods.R
\name{sigmavec2mat}
\alias{sigmavec2mat}
\title{sigmavec2mat}
\usage{
sigmavec2mat(sigma.vec, d)
}
\arguments{
\item{sigma.vec}{A vector of length d(d+1)/2 containing the lower triangle of a symmetric matrix}

\item{d}{The dimension of the matrix}
}
\value{
d by d symmetric matrix
}
\description{
Convert a vech (lower triangular, column-major) vector to a symmetric matrix
}
