% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_methods.R
\name{sigmavec2mat}
\alias{sigmavec2mat}
\title{sigmavec2mat}
\usage{
sigmavec2mat(sigma.vec, d)
}
\arguments{
\item{n}{The number of observations}

\item{alpha}{m by 1 vector that represents proportions of components}

\item{mu}{d by m matrix that represents mu}

\item{sigma}{d by d*m matrix that represents variance of components}
}
\value{
n by d vector
}
\description{
Convert sigma vector to matrix
}
